A rough implementation to handle events
Needed to get event handling in as quickly as possible in case Smitty decides to manually trigger one before the 3 days are up.

Will need some major refactoring when time permits, mostly around data storage. E.g. end of event info should be stored as a row in the configs table in the database.

Additionally, Discord webhooks are capped at 2000 characters so the final scores list will need to be split for posting to Discord.