estimated playtime chat stat
With a large enough sample size, the amount of playtime required for finding diamonds and blood diamonds can be considered to be constant. Therefore, the d/bd found stat can be used to estimate the amount of playtime accrued in the game over the tracked time period. This has been added to the dynamic chat command structure then handled like any other chat stat, including using the Chat.per_time() method for daily and hourly stats.